+++ util/sokol_imgui.h
@@ -2266,7 +2266,7 @@ static void _simgui_log(simgui_log_item_t log_item, uint32_t log_level, const ch
                 msg = _simgui_log_messages[log_item];
             }
         #endif
-        _simgui.desc.logger.func("simgui", log_level, log_item, msg, line_nr, filename, _simgui.desc.logger.user_data);
+        _simgui.desc.logger.func("simgui", log_level, (uint32_t)log_item, msg, line_nr, filename, _simgui.desc.logger.user_data);
     } else {
         // for log level PANIC it would be 'undefined behaviour' to continue
         if (log_level == 0) {
